Separate out recording from RecordingHiveMetastore
Split out the recording state from RecordingHiveMetastore, so recording
can be shared by multiple HiveMetastore instances.
